Create Decentralized_Bank.sol
The first step is to deploy the smart contract using the Remix IDE. After writing the code compile the code. When it is successfully compiled then deploy it. 
After deploying the contract a deployed Contract is obtained and then add the Amount as an owner. 
If Owner has to transfer, then add the receiver’s Address after that click on transfer.